+++ libs/game/scene.ts
@@ -30,44 +30,6 @@ namespace scene {
         lock: boolean;
         handler: () => void;
     }
-    
-    class Renderable implements SpriteLike {
-        protected _z: number;
-        public id: number
-
-        public constructor(
-            protected handler: (target: Image, camera: Camera) => void,
-            protected shouldBeVisible: () => boolean,
-            z: number,
-        ) {
-            this.z = z;
-
-            game.currentScene().addSprite(this);
-        }
-
-        get z(): number {
-            return this._z;
-        }
-
-        set z(v: number) {
-            if (this._z !== v) {
-                this._z = v;
-                game.currentScene().flags |= Flag.NeedsSorting;
-            }
-        }
-
-        get flags(): number {
-            return this.shouldBeVisible() ? sprites.Flag.None : sprites.Flag.Invisible;
-        }
-
-        __draw(camera: scene.Camera) {
-            this.handler(screen, camera);
-        }
-
-        __update() { }
-
-        __serialize(): Buffer { return undefined }
-    }
 
     export const CONTROLLER_PRIORITY = 8;
     export const TILEMAP_PRIORITY = 9;
@@ -221,25 +183,6 @@ namespace scene {
             this.collisionHandlers = undefined;
             this.gameForeverHandlers = undefined;
             this._data = undefined;
-        }
-
-        createRenderable(
-            z: number,
-            handler: (target: Image, camera: Camera) => void,
-            shouldBeVisible?: () => boolean
-        ): Renderable {
-            const renderable = new Renderable(
-                handler,
-                shouldBeVisible || (() => true),
-                z,
-            );
-
-            this.allSprites.push(renderable);
-            return renderable;
-        }
-
-        removeRenderable(renderable: Renderable) {
-            this.allSprites.removeElement(renderable);
         }
 
         protected cachedRender: Image;
